requirements.txt
pandas>=2.0
numpy>=1.24
openpyxl>=3.1
tqdm>=4.66
scikit-learn>=1.3
gensim>=4.3
tensorflow==2.15.1
tensorflow-metal>=1.1.0; platform_system == "Darwin" and platform_machine == "arm64"
google-api-python-client>=2.100
python-dotenv>=1.0
